Make client list function use the traffic-managers WatchWorkloads
Let the `telepresence list` command use a workload collection that is
backed by the traffic-managers `WatchWorkloads` function.

The client will use local shared informers to watch the workloads when
connecting to an older traffic-manager that doesn't support the
`WatchWorkloads` call.

Sort conditions by LastTransitionTime so that more recent has priority.
When checking the state of an argo-rollout, it's essential to look at
the transition timestamp, because the rollout can be both "progressing"
and "available" when pausing between steps.

We now look at the `RolloutAvailable` rather than `RolloutHealthy`. The
latter is not set until all steps of a canary rollout have completed,
which might be never.

Signed-off-by: Thomas Hallgren <thomas@tada.se>